getLocation

+++ src/controllers/locations-controller.ts
@@ -1,5 +1,6 @@
 import { RequestHandler } from "express";
 import { Error } from 'mongoose';
+import mongoose from 'mongoose';
 import LocationModel from "../models/location-model";
 
 type ErrorMessagesLT = {
@@ -34,6 +35,32 @@ export const getLocations: RequestHandler = async (req, res) => {
   res.status(200).json(locations);
 };
 
+export const getLocation: RequestHandler = async (req, res) => {
+  const { id } = req.params;
+  try {
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      res.status(400).json({
+        error: `Serverio klaida ieškant lokacijos: nevalidus id ${id}`,
+      });
+      return;
+    }
+    const location = await LocationModel.findById(id);
+    if (location) {
+       res.status(200).json({
+        location: location,
+      });
+    } else {
+      res.status(404).json({
+        error: `Lokacija su id ${id} nerasta`,
+      });
+    }
+  } catch (error) {
+    res.status(500).json({
+      error: error instanceof Error ? error.message : 'Serverio klaida ieškant lokacijos',
+    });
+  }
+}
+
 export const createLocation: RequestHandler = async (req, res) => {
   const locationProps = req.body;
   try {
@@ -54,16 +81,24 @@ export const deleteLocation: RequestHandler = async (req, res) => {
   const { id } = req.params;
 
   try {
-    const location = await LocationModel.findByIdAndDelete(id);
-    if (location === null) {
-      throw new Error(`Lokacija su id '${id}' nerasta`);
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      res.status(400).json({
+        error: `Serverio klaida ieškant lokacijos: nevalidus id ${id}`,
+      });
+      return;
     }
-
-    res.status(200).json({
-      location: location,
-    });
+    const location = await LocationModel.findByIdAndDelete(id);
+    if (location) {
+      res.status(200).json({
+        location: location,
+      });
+    } else {
+      res.status(404).json({
+        error: `Lokacija su id ${id} nerasta`,
+      });
+    }
   } catch (error) {
-    res.status(400).json({
+    res.status(500).json({
       error: error instanceof Error ? error.message : 'Serverio klaida trinant lokaciją',
     });
   }
